Move chr function to datafusion_functions
