+++ datafusion/expr/src/expr_fn.rs
@@ -584,12 +584,6 @@ scalar_expr!(
     string,
     "the number of characters in the `string`"
 );
-scalar_expr!(
-    Chr,
-    chr,
-    code_point,
-    "converts the Unicode code point to a UTF8 character"
-);
 scalar_expr!(InitCap, initcap, string, "converts the first letter of each word in `string` in uppercase and the remaining characters in lowercase");
 scalar_expr!(Left, left, string n, "returns the first `n` characters in the `string`");
 scalar_expr!(Reverse, reverse, string, "reverses the `string`");
@@ -1039,7 +1033,6 @@ mod test {
         test_scalar_expr!(Iszero, iszero, input);
 
         test_scalar_expr!(CharacterLength, character_length, string);
-        test_scalar_expr!(Chr, chr, string);
         test_scalar_expr!(Gcd, gcd, arg_1, arg_2);
         test_scalar_expr!(Lcm, lcm, arg_1, arg_2);
         test_scalar_expr!(InitCap, initcap, string);
